Fix the LTI config tab.
The LTIConfigvalues.config inclusion is moved to to the end of
defaults.config, instead of the end of authen_LTI.conf.  This means that
file is always included.  However, the config variables it defines are
only added to the `$ConfigValues` array if there are uncommented entries
in the `@LTIConfigVariables` array that is still defined in
authen_LTI.conf.  So the behavior is the same as before.