POD Viewer improvements
First, improve the efficiency of the WeBWorK::Utils::PODPaser package.

The $podFiles argument must be provided to the new method in order for
the module to resolve POD links to local files.  It should be the first
return value of the POD::Simple::Search survey method if called in list
context, or only return value if called in scalar context.

This is opposed to the previous method of search the root directory for
each link that was encountered in the POD.  Now instead, the directory
is searched for files containing POD only once, and the results are
stored in a hash.  This is done by the caller and passed to the
WeBWorK::Utils::PODPaser object on construction, instead of the
WeBWorK::Utils::PODPaser object handling it.

The module now expects `L<...>` usage to be correct according to the
documentation at https://perldoc.perl.org/perlpod#L-a-hyperlink.  Note
that it also now honors links to POD head sections in other local files.

The module also honors the following special cases that fit into that
scheme for local files:

* For a link to a local package, use the package name (exactly as the
  rules for `L<...>` dictate).
* For a link to a `.pod` file, just use the basename without the `.pod`
  extension.  So for example for a link to
  doc/MathObjects/UsingMathObjects.pod, use L<UsingMathObjects>.
* For a link to a macro file (or script), use the basename of the macro
  file with the `.pl` extension and no path.

The generate-ww-pg-pod.pl script now uses Mojo::Template to make it
easier to make html.

The POD page layouts have been improved with a sidebar for indices.
Mojo::DOM is used to parse the generate POD and pull the index out to
place in the sidebar.

Also add a layout template for the POD rendered by webwork2.