Fix hardcopy generation in the PG problem editor.
This just writes dummy empty TeX macros that are needed by the inclusion
of headandfoot.tex.  Then it overrides the headers set by that file.
Then it writes the hardcopySetHeader.tex file so that two column mode is
started if that theme is selected.  After writing the problem TeX, it
writes the hardcopySetFooter.tex file so that two column mode is
appropriately ended.  Since that file includes the copyright.tex file,
these files will now have the logo and copyright at the bottom.

All of the common files are also added to the zip file if TeX Source is
selected or if PDF file generation fails.