Make messages dismissible and PG editor tmp edits more clear
Any message that is added by the ContentGenerator addgoodmessage or
addbadmessage methods is now dismissible.  There is a close button to
the right that can be clicked to remove the message from the page.  In
additiona, a hidden button is also added to the end of the page.  If
there are any dismissible messages, then the button is shown by
javascript, and if clicked all of the dismissible messages are removed
from the page (and the button hidden again).  This is a convenience for
removing all of those messages at once if there are many of them.

In the PG editor, when a temporary file is created by javascript the
file name is changed to the temporary file name in the "Editing ..."
statement and the `temporaryFile` class is added which gives the text
the orange temporary file color.  Furthermore, the "Editing ..."
statement is now shown both above and below the codemirror editor in the
page.  This makes it more clear that a temporary file is being edited.

Also fix a bug introduced by #1918.  The PG editor revert form needs to
be shown for course information files as well.  Otherwise there is no
way to revert to the original file once a temporary file is created
other than saving the file.  @somiaj: What was your rationale for doing
this?

Also add site nav sub links for set headers and the course information
file.  Now that the problem editor is directly in the site nav that link
is shown as active when editing one of these files.  That is confusing
because if you click on that "active" link it does not open the page you
are currently viewing that is editing one of these files.  Instead it
opens the editor with a new problem.