Set the content type header in the correct place.
This only sets the "text/xml" header if the route is the "render_rpc"
route and the output format is "ptx" and the display mode is "PTX".
This is done in `FormatRenderedProblem` (instead of `RenderViaRPX` as I
suggested earlier) because it needs to be done before the response is
rendered, and that occurs at the end of the `formatRenderedProblem`
method.