Add the capability in the pg problem editor to pg perltidy code.
This adds a new tab to the PG problem editor (unless editing a course
configuration file).  The tab contains basic information about what it
does, but if the "Tidy Code" button is pressed it perltidies the code in
the CodeMirror editor window.  The tidied code is also saved to the
temprorary file.

The actual tidying is done by a new webwork webservice action (not a
form submission).  If errors occur, they are shown in the renderer
window.

The webwork webwservice depends on a new PG module defined in a
corresponding PG pull request (see https://github.com/openwebwork/pg/pull/868).