Fix download-OPL-metadata-release.pl to not need write permission to the cwd.
Currently the File::Fetch module downloads the release data JSON file
from GitHub and slurps its contents into a scalar.  When I implemented
this, I thought I was avoiding disk IO with this approach.  However,
File::Fetch is not that well implemented as it turns out.  What it is
actually doing is creating a temporary directory in the current working
directory, downloading the file into that directory, and then slurping
the contents.  This means that you must have write permission for
whatever directory you are running the script from in order for the
script to work.

This switches to downloading the file into the `$ce->{webworkDirs}{tmp}`
directory, then uses Mojo::File to slurp the contents of the downloaded
file. Once the file is read, then the file is deleted.  So basically,
this does the same thing, except it chooses which temporary directory is
used for the download.  You already need to have write permission for
`$ce->{webworkDirs}{tmp}` for the script to work, since that is where
the actual OPL metadata release is downloaded to.

See https://webwork.maa.org/moodle/mod/forum/discuss.php?d=8270 for the
problems caused by the current method.