Don't strip the password from the input parameters for the rpc endpoints.
Password authentication for the rpc endpoints relies on this, and this
is the cause of issue #2115.

There really is no reason that the password should ever be stripped from
the parameters.  I suspect someone at some point thought leaving it in
the parameters after authentication is complete was some sort of
security vulnerability.  It really isn't.  The parameters are secure in
the system.  Nothing outside of the application can see them.  In fact
no real security is even gained by removing this parameter either.  So
the line modified here could simply be safely deleted instead.  For now,
I left and only leave the parameter for the rpc endpoints though.