Fix codemirror folding for html and xml.
Currently when you edit a course info file in the PG problem editor,
there is a console log error.  This is because the PG code folding
options do not work for htmlmixed mode.  This will also happen with the
xml mode that will be added in pull request #2114.

This uses different options for htmlmixed or xml mode than for PG mode
so that folding works for html and xml as well.

This also only enables comment toggling for PG mode.  That doesn't work
for html and xml.