Correct htmlPath and imagesPath settings in defaults.config
These paths have been incorrect since 2015 when they were added.  The
comment that was there about these is completely incorrect.  It states
that these should be complete urls.  They should not be urls at all.
Instead they should be directories.  The only place that these are used
is in PGalias.pm to search for auxiliary html files or images, and there
they are treated as directories.

Also remove the completely unused pdfPath setting.