Possible fix for CAS.pm authentication issue.
This replaces the mod_perl `$r->unparsed_uri` method with
Mojo::Message::Request method `$c->req->url->to_string` which should
return the same thing.  This may fix the `CAS.pm` module.  This is the
only mod_perl call in that file, and was missed in the conversion to
Mojolicious.

I have not way to test this, but clearly that method must be replaced.

See https://webwork.maa.org/moodle/mod/forum/discuss.php?d=8317.