Fix the PG Problem Editor buttons above the editor window.
The `if` statement checking for the PGML labs problem went to far, and
included the "Author Info" and "Report Bugs in this Problem" buttons,
and the ending tag for the div that contains all of the buttons.

Those buttons are not present when editing the course information file.

This may have been caused by a merge conflict resolution issue.