Switch the CourseEnvironment and AchievementEvaluator back to WWSafe.
It turns out that using the default Safe is causing problems still.
When a problem fails to render within the (hardcoded default) 60 second
`$WeBWorK::PG::TIMEOUT`, then the alarm signal is not caught by the
`$SIG{ALRM}` handler that is set in `WeBWorK::PG`.  As a result, instead
of getting the information about the problem failing to render within
the timeout, you instead get a useless message about decode_json failing
in Mojo::JSON.

I suspect this is also the cause of issues #2041 and #2148 (now closed).
Although the cause of those issues is really an error in the problems,
this was hiding the error in the problems (combined with the issue fixed
in pull request https://github.com/openwebwork/pg/pull/883 for issue #2041).

However, this is not PG's WWSafe module.  This is webworks own
WeBWorK::WWSafe module.  This is used to avoid conflict with PG's
module, and not depend on PG's library for scripts.