Tweak the site navigation link colors.
This makes the background color lighter and the foreground color dark
for active links in the the site navigation.  This makes it easier to
determine when the active link has focus because the outline color is
darker than the new background color.

The math4-green theme needed a slightly darker background than the
math4 and math4-red themes, and the the colors are the same as before
for the math4-yellow theme for which this is not an issue.

This addresses the first part of issue #2157.