Progress bar accessibility improvements.
The progress bar elements are now focusable, and tooltips active on both
hover and focus.  Note that this makes all `.set-id-tooltips` active on
hover and focus.  That means that this also applies to the actual set
ids as well on the problem sets page.

The box shadows on the progress bar and contained elements are now
black and thicker.

When a progress bar element is focused the border is thicker and box
shadow a bit lighter.  This may not be good enough by itself, but should
be good enough considering that the tooltip is really what is important
here.  In fact I think that visual aspects of the progress bar in
general are not as important with the tooltip being active on focus.

Also make the style more specific.  The `.progress` and `.progress-bar`
classes are bootstrap classes that could conceivably be used for another
progress bar.  So a `.set-progress` class is added to distinquish this
from other progress bars.