Improvements to the course directory update method.
This always creates a course directory if it is missing.  It also
attempts to fix permissions if those are not correct.  It does not
attempt to change ownership or the group as that will almost always
fail, and will need to be done manually.

The directories that can be copied from the model course are copied if a
directory is created.  This is only done if the path did not exist to
begin with.