Move the ConfigValues array out of defaults.config.
This array is now defined in `lib/WeBWorK/ConfigValues.pm` and is
returned by the `getConfigValues` method.  The `$LTIConfigValues` hash
is also defined in that file and added to the array when
`getConfigValues` is called if LTI authentication is enabled for the
course.  This means that these are no longer configurable.  No config
file can modify them. Furthermore, the array is not even ever added to
the course environment.  Instead it is just used by the course
configuration module which is the only place it was ever used.

This means that if a course's course.conf file includes authen_LTI.conf,
it does not also need to include the LTIConfigValues.config file.  That
file in fact no longer exists.

Also remove all access to the `paramcache` outside of the
`WeBWorK::Controller` module.  That is an internal implementation detail
that should never be accessed directly outside of that file.