Rewrite the dump_past_answers script (renamed to dump-past-answers.pl).
The script no longer saves ids (course id, set id, user id, etc)  as
useless hashes.  I don't know why it was thought that doing so was a
good idea.

The script does not gzip the resulting csv file.  You can do that on
your own if so desired.

The script does not have an upload option.  You can upload it to where
you want on your own if so desired.

The unix timestamp for the time at which the script is executed is no
longer saved in each row of the CSV file.  It is part of the default
file name if you want it.

Past answers from gateway quiz versions are also dumped.  I think it was
intended that they would be dumped before, but it wasn't working.

Past answers from problems with more than one answer in them are now
dumped.  It certainly was intended that this be done before, but wasn't
because of a bug in the script.