Tweaks to the file manager make archive page.
If the chose archvie file exists, then don't blindly overwrite it.  Add
a checkbox to that allows the user to select to overwrite the file.
Only replace the existing archive file if that checkbox is checked.
Otherwise show a message that the file exists, and don't overwrite.

If a single file is selected (or single directory initially selected)
then use that filename (without extension) for a default for the archive
name.