Improved file manager button states.
The buttons that can only act on one file are now disabled when multiple
files are selected. Only the "Delete" and "Make Archive" buttons can act
on multiples files.

The type of file (regular, directory, or link) is stored in a data
attribute for each option in the `files` select.  The javascript also
takes this into consideration and is disabled for a file type a button
can not act on.  The "View", "Edit", "Download", and "Copy" buttons can
not act on symbolic links.  The "Edit", "Download", and "Copy" buttons
can not act on directories.