Change the way that the archive type is determined.
This adds a "By extension" option to the archive type dropdown which is
the default option.  If that option is selected, then the archive type
will be determined by the extension given in the file name.  If the file
name does not have a valid archive extension, then the zip archive type
is assumed and that extension is added.  If zip or tar is selected then
the zip or tgz extension is added to the filename (if it doesn't already
have that extension), and of course that archive type is used. Note that
describes the server side behavior.  Javascript also changes the
extension in the input client side.

Note that the `.tar.gz` extension is also supported if explicitly given
for the filename extension.

Note that at or above the large breakpoint the archive type dropdown and
"Overwrite existing archive" checkbox will be on the second line
together.  Below the large breakpoint, they will be on separate lines.

Also `dir="ltr"` is added to both the archive filename input and the
files select.

Also remove the "Create archive of the selected files?" question.  The
presented answers are not valid for that question, and the question
isn't needed with the instructions at the top, as well as the fact that
the buttons are clear as to what they will do.