Switch to using `Archive::Zip` and `Archive::Tar` to extract archives.
The `Archive::Extract` module is not safe with zip files and can extract
files outside of the current working directory (assuming the server has
write permission to do so).

The `Archive::Zip` and `Archive::Tar` modules will not extract outside
of the current working directory (by default).

Additionally, using the `Archive::Zip` and `Archive::Tar` modules gives
more power for archive extraction. If the "Overwrite existing files
silenetly" checkbox is not checked, then this now checks each file in
the archive to see if extracting it will overwrite an existing file.  If
so it refuses to do so, and a message is displayed.