Fix messages for existing files or files outside the course directory.
This makes it so that the first thirty files in both lists are shown.
If there are more than thirty, then the last item in the list will say
that there are more files not shown.

To make this work a `min` method was added to Utils.pm (there was a max
method, but no min method?), and the change in #2194 to make good and
bad messages be in a `div` instead of a `p` was added here too.