Remove the `usePolling` chokidar option in generate-assets.js.
With node version 20 the `usePolling` option is resulting in high cpu
usage.  The `followSymlinks` serves the purpose more correctly.  The
`usePolling` option was initially used because either the
`followSymlinks` option did not exist or there was a bug in chokidar and
that option wasn't working.  I can't remember the exact details. Note
that `followSymlinks` is true by default, and so it is not explicitly
set.

Also remove the `interval` option since that is only used when
`usePolling` is true.

The actual high cpu usage occurs with the PG generate-assets.js script,
and not this one with the `usePolling` option.  I don't really know why
though.  In any case the `followSymlinks` option is more correct for
what we want here.