Watch the PG.pl macro file when running morbo.
Since this file is cached in the safe zone when the application starts,
changes to this file require the webwork2 application to be restarted in
order for them to take effect.  I have noticed this recently when making
changes to that file.

This of course only affects running the webwork2 app in development with
morbo.  With hypnotoad you would need to restart the app manually since
hypnotoad doesn't hot reload.  Of course you normally don't edit PG.pl
in production, so that is how it should be.