Fix `maketext` interpolation in PG.
The key to making this work is to override the `_compile` method of
`Locale::Maketext` and remove the `use strict` usage inside of its code
eval call. That is the only essential difference between the override
method added here and the orignal method in `Locale::Maketext`.

With this you can now use maketext interpolation in PG. For example, you
can call `maketext('[quant,_1,dog] ate my homework', $num_dogs)`.

This is an important part of removing the attempts table, and moving the
code the provides feedback to PG.

The WeBWorK::Localize module is also cleaned up considerably.