Add a page to manage jobs in the job queue.
All jobs for a course are listed on this page. The table displays the
job id (this is used to reference specific jobs in action messages,
otherwise it would not be shown), task name, created time, started time,
finished time, and state.  Also a button that opens a popover containing
the job result is in the state column if the job has completed.

Note that the Minion job queue automatically removes jobs from the job
queue after two days (that is the default at least which we don't
change).  So the real importance of this page is to allow the instructor
to see the status of recently completed or in progress jobs.

At this point the actions available on the page are filter, sort, and
delete.  Jobs can be filtered by id, task name, or state. Jobs can be
sorted by clicking on the headers, or by using the sort form.  Jobs that
are not active can be deleted.

Minion does not allow deletion of active jobs.  Note that an active job
means a job that is currently running. As such they can not be selected
on this page. Perhaps an option to stop running jobs could be added at
if there is a problem with jobs hanging, but active jobs can not be
directly stopped. The Minion worker is in a different process so the
Mojolicious app needs to broadcast a signal to the Minion worker to do
so.

An inactive job (i.e., a job that has been queued but has not started
running yet) can be selected and deleted.  However, it is possible that
the inactive job could start before the form is submitted.  In that case
the job can not be deleted, and so an alert will show that.

In order to reliably associate a course with a job there is a new rule
for tasks.  The job must pass the course id via the "notes" option of
the Minion enqueue method.  The existing tasks have been updated to do
this. There is also a backwards compatibility check to find jobs that
passed it one of the ways the two jobs did it before in the job
arguments.

Since the job fail/finish messages are now displayed in the UI, those
messages are now translated. That is all except the first few messages
in each task before the course environment is established, since a
course environment is required to obtain the language of the course.

The send_instructor_email task no longer sends an email to the
instructor after sending the emails to the students.  Instead the job
result contains all of the information that would have been in that
email. This is a far more reliable way of getting that information to
the instructor sending the email.  The instructor just needs to go to
the "Job Manager" page to see the result.  The message on the "Email"
page tells the instructor this.

This page is also available for the admin course.  In the admin course
all jobs for all courses are shown.  There is an additional column in
the jobs table that shows the course id for the course the job was
enqueued by.

The errors that are reported when sending emails are made less verbose
by calling the `message` method of the Mojo::Exception which does not
include the traceback.