Add sort by column header click to the homework sets editor.
This uses the same method as the classlist editor for this.  In fact the
javascript used by the classlist editor for this is now in the
actiontabs.js file, and is used by both the classlist editor and the
homework sets editor.

The column headers in the homework sets editor did not make sense.  The
headers should say what is in the column, not what clicking on an extra
item in the column does.  This was made worse by making it so that
clicking on the column header sorts by set name.  It doesn't make sends
to click on "Edit Set Data" to sort the column by set name.  Other
headers also were adjusted to remove the "Edit" wording, and reflect
contents not actions.  We need to find a better way to convey what
clicking on something in the column does.  A tooltip is added to the
pencil icon by the set name to convey what it does.

Using the set name header as a label for the select all is problematic.
So instead a double check icon is added and a tooltip on the select all
input itself.  This is also done in the other tables like this (the
classlist editor and achievement list).

Remove the "Please select action to be performed" messages that were
displayed when the homework sets editor, classlist editor, or
achievement editor pages were initially loaded.  Those messages aren't
needed.  Furthermore, remove the "Result of last action performed:"
prefix that was added to all action messages.  That doesn't add anything
useful to the message, and the action messages are clear on their own.
The messages were improved with capitalization and periods to make them
look better when shown without that prefix.