Add reverse sorting
This is implemented for both the homework sets editor and the classlist
editor.

The sort form now has an additional order dropdown for each of the sort
priorities to choose if the sort is ascending or descending for that
field.

If the header of the primary sort field in the table is clicked, then
the only change in the sorting is that the order (ascending or
descending) of the primary sort is toggled.

There is also a button added in each of the table headers corresponding
to the fields used for sorting.  The button shows a number (1, 2, or 3)
indication the sort priority of that field, and a chevron up or down
indicating if the sort order is descending or ascending, respectively.
Clicking on that button sorts again, but only toggles if the sort for
that priority level is ascending or descending.