Capture logs for the LTIMassUpdate task and return them for the job result.
Note that the job will still succeed but now some failures/messages will
appear in the job result.  If `debug_lti_grade_passback` or
`debug_lti_parameters` are enabled then the job result will now be
rather extensive.  Everything that is sent to the log will be in result.

Also add logs back that were removed initially.  Since Minion removes
jobs after two days, the messages in the result are lost with them.  So
also log them so that they are more permanently stored.