Update third party JavaScript dependencies.
The following dependencies were updated.

* @fortawesome/fontawesome-free from 6.2.1 to 6.4.2
* bootstrap from 5.2.3 to 5.3.2
* codemirror from 5.65.11 to 5.65.13
* iframe-resizer from 4.3.2 to 4.3.7
* jquery from 3.6.3 to 3.7.1
* luxon from 3.2.1 to 3.4.3
* autoprefixer from 10.4.13 to 10.4.16
* cssnano from 6.0.0 to 6.0.1
* rtlcss from 4.0.0 to 4.1.1
* sass from 1.57.1 to 1.69.5
* terser from 5.16.1 to 5.24.0
* yargs from 17.6.2 to 17.7.2

These versions of these libraries are all compatible with webwork's
current usage and require no changes to the code except bootstrap. That
requires a few minor changes.

All of the dependencies are now at their newest versions except for
codemirror which will need some work to update.

Note that the new version of bootstrap brings lots of improvements that
could be beneficial.  In particular it has color modes.  It will take
some work to implement (particularly with addressing accessibility
concerns), but a dark mode could now be added.